linuxmuster-base (1.2-10) testing; urgency=low
  * minor optimations in wimport-linbo.sh and import_workstations
  * fixed lockfile handling in samba-userlog
  * minor optimations in linuxmuster-config and linuxmuster-patch
    concerning to imaging system configuration and installation

linuxmuster-base (1.2-9) testing; urgency=low

  * added linuxmuster-imaging-tivoli to package dependencies
  * added tpmfos configuration support to linuxmuster-setup
  * added webcache port (8080) to base-updates/allowed_ports

linuxmuster-base (1.2-8) testing; urgency=low

  * added deactivation of wwwadmin's samba account to repair-admins.sh
  * improved lockfile handling in samba-userlog

M    share/scripts/linuxmuster-patch
M    share/scripts/wimport_linbo.sh
M    share/scripts/linuxmuster-config
M    share/scripts/repair-admins.sh
M    debian/control
M    debian/postinst
M    debian/changelog
M    debian/templates
A    var/config-dynamic/46_tpmfos
A    var/config-dynamic/46_tpmfos/config.csv
A    var/config-dynamic/46_tpmfos/config.csv.target
A    var/config-dynamic/46_tpmfos/postpatch
A    var/config-dynamic/46_tpmfos/prepatch
M    var/config-dynamic/00_stop-services/prepatch
A    var/config-dynamic/03_dhcp3-server/dhcpd.conf.tivoli
M    var/config-dynamic/45_rembo/prepatch
AM   var/config-static/etc/init.d/tpmfos-derby
M    var/base-updates/allowed_ports
M    sbin/samba-userlog
M    sbin/import_workstations

+++ share/scripts/wimport_linbo.sh
@@ -23,7 +23,7 @@ check_unique() {
 exitmsg() {
 	echo "$1"
 	rm $WDATATMP
-	rm $locker
+	rm -f $locker
 	exit 1
 }
 
@@ -216,17 +216,14 @@ if [ -s "$WIMPORTDATA" ]; then
 		echo "}" >> $DHCPDCONF
 
 		# linbo stuff, only if pxe host
-		if [ "$pxe" != "0" ]; then
-			if [ "$linbo" = "yes" ]; then
-				if [ ! -e "$LINBODIR/start.conf.$hostgroup" ]; then
-					echo "  * LINBO: Creating new configuration for hostgroup $hostgroup ..."
-					cp $LINBODEFAULTCONF $LINBODIR/start.conf.$hostgroup
-				fi
-				echo "  * LINBO: Linking IP $ip to hostgroup $hostgroup ..."
-				[ -e "$LINBODIR/start.conf-$ip" ] && rm -rf $LINBODIR/start.conf-$ip
-				ln -sf start.conf.$hostgroup $LINBODIR/start.conf-$ip
-			else
-				echo "  * Skipping LINBO configuration"
+		if [[ "$pxe" != "0" && "$linbo" = "yes" && "$imaging" = "linbo" ]]; then
+			if [ ! -e "$LINBODIR/start.conf.$hostgroup" ]; then
+				echo "  * LINBO: Creating new configuration for hostgroup $hostgroup ..."
+				cp $LINBODEFAULTCONF $LINBODIR/start.conf.$hostgroup
+			fi
+			echo "  * LINBO: Linking IP $ip to hostgroup $hostgroup ..."
+			[ -e "$LINBODIR/start.conf-$ip" ] && rm -rf $LINBODIR/start.conf-$ip
+			ln -sf start.conf.$hostgroup $LINBODIR/start.conf-$ip
 			fi
 		fi
